pyproject.toml
[project]
name = "livekit-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "agentmail>=0.0.45",
  "agentmail-toolkit>=0.1.29",
  "asyncio>=3.4.3",
  "livekit-agents[deepgram,openai,cartesia,silero,turn-detector]~=1.0",
  "livekit-plugins-noise-cancellation~=0.2",
  "python-dotenv",
  "websockets>=15.0.1",
]
